Refactor volumes and labels
- Move labels to their own file as they just create noise and are very simple
- Move volumes to their own file and make an interface for them. We'll
  expand on this in the future to support user mounts.
